+++ scripts/gewitter_functions.py
@@ -1,3 +1,12 @@
+""" 
+Description:
+-----------
+This script hosts many helper functions to make notebooks cleaner. The hope is to not distract users with ugly code. 
+
+Alot of these were sourced from Dr. Lagerquist and found originally in his gewitter repo (https://github.com/thunderhoser/GewitterGefahr).
+
+"""
+
 #additional libraries needed here
 import scipy.stats as st 
 import copy 
@@ -414,201 +423,6 @@ pretty_names = [ '$\lambda_{\downarrow}$', '$T_{d}$', '$V_{fric}$', 'Gflux', '$C
  '$T_{sfc}$ - $T_{2m}$']
 
 
-import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
-
-def percentile(n):
-    def percentile_(x):
-        return np.percentile(x, n)
-    percentile_.__name__ = 'percentile_%s' % n
-    return percentile_
-
-#determine unconditional mean, sum R in each bin. But then devide by master counts
-def boxbin(x,y,xedge,yedge,c=None,figsize=(5,5),cmap='viridis',mincnt=10,vmin=None,vmax=None,edgecolor=None,powernorm=False,
-           ax=None,normed=False,method='mean',quantile=None,alpha=1.0,cbar=True,unconditional=False,master_count=np.array([])):
-    
-    """ This function will grid data for you and provide the counts if no variable c is given, or the median if 
-    a variable c is given. In the future I will add functionallity to do the median, and possibly quantiles. 
-    
-    x: 1-D array 
-    y: 1-D array 
-    xedge: 1-D array for xbins 
-    yedge: 1-D array for ybins
-    
-    c: 1-D array, same len as x and y 
-    
-    returns
-    
-    axis handle 
-    cbar handle 
-    C matrix (counts or median values in bin)
-    
-    """
-    
-    midpoints = np.empty(xedge.shape[0]-1)
-    for i in np.arange(1,xedge.shape[0]):
-        midpoints[i-1] = xedge[i-1] + (np.abs(xedge[i] - xedge[i-1]))/2.
-    
-    #note on digitize. bin 0 is outside to the left of the bins, bin -1 is outside to the right
-    ind1 = np.digitize(x,bins = xedge) #inds of x in each bin
-    ind2 = np.digitize(y,bins = yedge) #inds of y in each bin
-    
-    
-    #drop points outside range 
-    outsideleft = np.where(ind1 != 0)
-    ind1 = ind1[outsideleft]
-    ind2 = ind2[outsideleft]
-    if c is None:
-        pass
-    else:
-        c = c[outsideleft]
-        
-    outsideright = np.where(ind1 != len(xedge))
-    ind1 = ind1[outsideright]
-    ind2 = ind2[outsideright]
-    if c is None:
-        pass
-    else:
-        c = c[outsideright]
-        
-    outsideleft = np.where(ind2 != 0)
-    ind1 = ind1[outsideleft]
-    ind2 = ind2[outsideleft]
-    if c is None:
-        pass
-    else:
-        c = c[outsideleft]
-    outsideright = np.where(ind2 != len(yedge))
-    ind1 = ind1[outsideright]
-    ind2 = ind2[outsideright]
-    if c is None:
-        pass
-    else:
-        c = c[outsideright]
-    
-
-    if c is None:
-        c = np.zeros(len(ind1))
-        df = pd.DataFrame({'x':ind1-1,'y':ind2-1,'c':c})
-        df2 = df.groupby(["x","y"]).count()
-        df = df2.where(df2.values >= mincnt).dropna()
-        C = np.ones([xedge.shape[0]-1,yedge.shape[0]-1])*-9999
-        for i,ii in enumerate(df.index.values):
-            C[ii[0],ii[1]] = df.c.values[i]
-        C = np.ma.masked_where(C == -9999,C)
-        
-        if normed:
-            n_samples = np.ma.sum(C)
-            C = C/n_samples
-            C = C*100
-            print('n_samples= {}'.format(n_samples))
-        
-        if ax is None:
-            fig = plt.figure(figsize=(5,5))
-            ax = plt.gca()
-        else:
-            pass
-            
-        if powernorm:
-            pm = ax.pcolormesh(xedge,yedge,C.transpose(),cmap=cmap,edgecolor=edgecolor,norm=colors.PowerNorm(gamma=0.5),vmin=vmin,vmax=vmax,alpha=alpha)
-            
-            if cbar:
-                cbar = plt.colorbar(pm,ax=ax)
-            else:
-                cbar = pm 
-        else:
-            pm = ax.pcolormesh(xedge,yedge,C.transpose(),cmap=cmap,vmin=vmin,vmax=vmax,edgecolor=edgecolor,alpha=alpha)
-            if cbar:
-                cbar = plt.colorbar(pm,ax=ax)
-            else:
-                cbar = pm 
-            
-        return ax,cbar,C
-    
-    elif unconditional:
-    
-        df = pd.DataFrame({'x':ind1-1,'y':ind2-1,'c':c})
-        if method=='mean':
-            df2 = df.groupby(["x","y"])['c'].sum()
-            
-        df3 = df.groupby(["x","y"]).count()
-        df2 = df2.to_frame()
-        df2.insert(1,'Count',df3.values)
-        df = df2.where(df2.Count >= mincnt).dropna()
-        C = np.ones([xedge.shape[0]-1,yedge.shape[0]-1])
-        for i,ii in enumerate(df.index.values):
-            C[ii[0],ii[1]] = df.c.values[i]
-                
-        C = C/master_count.values
-
-        if ax is None:
-            fig = plt.figure(figsize=(5,5))
-            ax = plt.gca()
-        else:
-            pass
-        
-        if powernorm:
-            pm = ax.pcolor(xedge,yedge,C.transpose(),cmap=cmap,vmin=vmin,vmax=vmax,norm=colors.PowerNorm(gamma=0.5),alpha=alpha)
-            if cbar:
-                cbar = plt.colorbar(pm,ax=ax)
-        else:
-            
-            pm = ax.pcolor(xedge,yedge,C.transpose(),cmap=cmap,vmin=vmin,vmax=vmax,alpha=alpha)
-            if cbar: 
-                cbar = plt.colorbar(pm,ax=ax)
-        
-        
-    else:
-        df = pd.DataFrame({'x':ind1-1,'y':ind2-1,'c':c})
-        if method=='mean':
-            df2 = df.groupby(["x","y"])['c'].mean()
-        elif method=='std':
-            df2 = df.groupby(["x","y"])['c'].std()
-        elif method=='median':
-            df2 = df.groupby(["x","y"])['c'].median()
-        elif method=='qunatile':
-            if quantile is None:
-                print('No quantile given, defaulting to median')
-                quantile = 0.5
-            else:
-                pass
-            df2 = df.groupby(["x","y"])['c'].apply(percentile(quantile*100))
-            
-            
-        df3 = df.groupby(["x","y"]).count()
-        df2 = df2.to_frame()
-        df2.insert(1,'Count',df3.values)
-        df = df2.where(df2.Count >= mincnt).dropna()
-        C = np.ones([xedge.shape[0]-1,yedge.shape[0]-1])*-9999
-        for i,ii in enumerate(df.index.values):
-            C[ii[0],ii[1]] = df.c.values[i]
-
-        C = np.ma.masked_where(C == -9999,C)
-
-        if ax is None:
-            fig = plt.figure(figsize=(5,5))
-            ax = plt.gca()
-        else:
-            pass
-        
-        if powernorm:
-            pm = ax.pcolor(xedge,yedge,C.transpose(),cmap=cmap,vmin=vmin,vmax=vmax,norm=colors.PowerNorm(gamma=0.5),alpha=alpha)
-            if cbar:
-                cbar = plt.colorbar(pm,ax=ax)
-            else:
-                cbar = pm
-        else:
-            
-            pm = ax.pcolor(xedge,yedge,C.transpose(),cmap=cmap,vmin=vmin,vmax=vmax,alpha=alpha)
-            if cbar: 
-                cbar = plt.colorbar(pm,ax=ax)
-            else:
-                cbar = pm 
-            
-    return ax,cbar,C
-
-
 def make_performance_diagram_axis(ax=None,figsize=(5,5),CSIBOOL=True,FBBOOL=True,csi_cmap='Greys_r'):
     import matplotlib.patheffects as path_effects
     pe = [path_effects.withStroke(linewidth=2,
@@ -637,7 +451,8 @@ def make_performance_diagram_axis(ax=None,figsize=(5,5),CSIBOOL=True,FBBOOL=True
     
     ax.set_xlim([0,1])
     ax.set_ylim([0,1])
-    
+    ax.set_xlabel('SR')
+    ax.set_ylabel('POD')
     return ax
 
 def get_mae(y,yhat):
